Fixed trying to build ddb for client

+++ src/client/serialize/daos_serialize.c
@@ -637,7 +637,7 @@ deserialize_roots(hid_t file_id, struct daos_prop_entry *entry, const char *prop
 		D_GOTO(out, rc = -DER_MISC);
 	}
 	attr_dtype_size = H5Tget_size(attr_dtype);
-	if (attr_dtype_size < 0) {
+	if ((int)attr_dtype_size < 0) {
 		D_ERROR("failed to get attribute type size\n");
 		D_GOTO(out, rc = -DER_MISC);
 	}
